requirements.txt
torch>=2.0
numpy>=1.24,<2.0
scipy>=1.10
pandas>=1.5
scanpy>=1.9
anndata>=0.9
h5py>=3.8
statsmodels>=0.14
scikit-learn>=1.2
matplotlib>=3.7
seaborn>=0.12
networkx>=3.0
leidenalg>=0.9
python-igraph>=0.10
transformers>=4.30
umap-learn>=0.5
